requirements.txt
colorama>=0.4.6
pyyaml>=6.0